+++ site/src/components/VppCurves.svelte
